fix: shuffled warp pads proximity detection
Since the warp pads in gem stone valley are placed so closely to each
other, they have a shorter proximity trigger for displaying the level
name. Originally, this trigger was based on level ID, not on the hub
the warp pad was in. This wouldnt do with our warp pad shuffle.